feat(astro): add astro integration
