+++ astro.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "astro/config";
+import react from "@astrojs/react";
+
+export default defineConfig({
+  integrations: [react()],
+  outDir: "./out", // aws amplify
+});
