--- pages/_app.tsx
@@ -1,5 +1,0 @@
-import { AppProps } from "next/app";
-
-export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
