--- pages/_document.tsx
@@ -1,15 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from "next/document";
-
-class MyDocument extends Document {
-  render() {
-    return (
-      <Html lang="en">
-        <Head />
-        <Main />
-        <NextScript />
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
