+++ src/pages/about.astro
@@ -1,0 +1,8 @@
+---
+import About from '../components/about.tsx';
+import Layout from '../layouts/layout.tsx';
+---
+
+<Layout title="About Jeff Mendez and personal life" description="A section about myself day to day and as a software engineer.">
+  <About client:visible />
+</Layout>
