+++ src/pages/hire.astro
@@ -1,0 +1,8 @@
+---
+import Hire from '../components/hire.tsx';
+import Layout from '../layouts/layout.tsx';
+---
+
+<Layout>
+  <Hire client:visible />
+</Layout>
