+++ src/pages/index.astro
@@ -1,0 +1,8 @@
+---
+import Home from '../components/home.tsx';
+import Layout from '../layouts/layout.tsx';
+---
+
+<Layout home description="Florida based software engineer focusing on accessibility at A11yWatch. Lover of art and music and of course coffee.">
+  <Home client:visible />
+</Layout>
