+++ src/pages/tags.astro
@@ -1,0 +1,8 @@
+---
+import Tags from '../components/tags.tsx';
+import Layout from '../layouts/layout.tsx';
+---
+
+<Layout title="Software Projects" description="Some of the projects I have made or contributed to.">
+  <Tags client:visible />
+</Layout>
